Remove Winston Intake
