+++ discord_bot/Dockerfile
@@ -18,5 +18,4 @@ RUN apt-get update && \
     rm -rf /var/lib/apt/lists/*
 RUN /usr/sbin/update-ca-certificates
 COPY --from=builder /usr/local/cargo/bin/discord_bot /usr/local/bin/discord_bot
-COPY --from=builder /usr/src/app/ai /usr/local/bin/ai
 CMD ["discord_bot"]
