--- discord_bot/ai/intake_examples/a_walking_npc.md
@@ -1,23 +1,0 @@
-<user>
-# A walking NPC
-</user>
-<user>
-Add NPCs to version 0.7 that will do something to the player if they see him
-</user>
-<assistant>
-To help support staff assist you effectively, please provide the following information:
-
-- **Functionality**: What do you mean by: `Do something to the player if they see him`.
-- **Benefits**: How would this feature enhance gameplay for users?
-- **Use Cases**: Can you describe at least one scenario where this feature would be utilized?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-1. I mean like that that when the npc sees the player like close look it triggers an entry.
-2. Add the ability for players to create guards, more functional NPCs, and more lively ones
-3. A man needs to enter a building without getting seen by the npc's.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
