--- discord_bot/ai/intake_examples/add_the_dialogue_function_using_actionbar.md
@@ -1,29 +1,0 @@
-<user>
-# Add the dialogue function using Actionbar
-</user>
-<user>
-Can we add Actionbar-based dialogue and selection functions? Due to the limited space of Chat, in some cases, the option function occupies part of the space, so it is difficult to fully display the content expressed.
-</user>
-<assistant>
-To get a support person to help you better, please provide the following information:
- - **Could you describe the problem clearly?** Maybe make a drawing or annotate a screenshot? Specify what is in this screenshot and what you want to achieve.
- - **How would other users usse this feature?** What are the use cases for this feature?
- - **What would be the advantages of this feature?** Why should this feature be added? How would it enhance the user experience?
-
- Providing the above information is **mandatory** for the ticket to be answered and look by the support team.
-</assistant>
-<user>
-Currently, things like option dialogue can only be displayed in the chat. But I want to display it in the action bar.
-Like have the conversation and option displayed in the action bar.
-</user>
-<assistant>
-- What would be the advantages of this feature? 
-- Why should this feature be added?
-</assistant>
-<user>
-I think it is easier to read and understand the conversation and options in the action bar rather than in the chat.
-It is more central and can look cooler.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
