--- discord_bot/ai/intake_examples/apex_host_support.md
@@ -1,32 +1,0 @@
-<user>
-# Typewriter - Apex Host Support?
-</user>
-<user>
-Hello it's been a while that I have created this support: https://discord.com/channels/1054708062520360960/1208744946002165792 it was around February right? then my last reponse was April 20, 2024;
-
-Now as I can see here it says in Beta and In Production, does it mean that it now support ApexHosting and this plugin will work?
-</user>
-<assistant>
-To assist you effectively, please provide the following information:
-
-- **Opening ports** Are you able to open 2 additional ports on Apex Hosting?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-I have no idea what that means, I'm not a developer, I just want to use it
-</user>
-<assistant>
-Typewriter needs 2 fresh network ports to work.
-Some hosting providers allow you to open additional ports, but others do not.
-
-Network ports are essentially identifiers for processes in Typewriter.
-
-Are you able to open 2 additional ports on Apex Hosting?
-</assistant>
-<user>
-No, I can't open any more ports
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
