--- discord_bot/ai/intake_examples/ender_dragon_entity.md
@@ -1,21 +1,0 @@
-<user>
-# Add Ender Dragon
-</user>
-<user>
-Please add the ender dragon entity.
-</user>
-<assistant>
-To help support staff assist you effectively, please provide the following information:
-
-- **Did you check?** Has the Ender Dragon entity not already been added to the entity extension?
-- **Does this have priority?** Should we add the addition of this entity to our priority list?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-1. Yes this entity hasn't been added to the entity extension
-2. No take your time./Yes i would like to have this get added as soon as possible.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
