--- discord_bot/ai/intake_examples/error_in_path_calculation.md
@@ -1,29 +1,0 @@
-<user>
-# Error in path calculation
-</user>
-<user>
-Does anyone know why this error may be occurring?
-</user>
-<assistant>
-To allow a support person to help you more effectively, please provide the following information:
-
-- **What were you trying to do when the error occurred?** Outline the steps you took leading up to the error.
-- **Have you checked the logs?** If so, could you upload the `latest.log` from your `logs` folder to [McLogs](https://mclo.gs) and share the link?
-
-Providing the above information is **mandatory** for the ticket to be answered and look by the support team.
-</assistant>
-<user>
-I was trying to use the PathStreamEntry to show a path to the players.
-Then I got the error in the file.
-</user>
-<assistant>
-Please provide the full logs from your `logs` folder. We gather a lot of information about the context of your server from the logs.
-Please upload the `latest.log` to [McLogs](https://mclo.gs) and share the link.
-</assistant>
-<user>
-Logs:
-https://mclo.gs/y4NrASr
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
