--- discord_bot/ai/intake_examples/groups_of_custom_sounds.md
@@ -1,41 +1,0 @@
-<user>
-# Groups of Custom Sounds
-</user>
-<user>
-Is there any way to create groups of custom sounds? (e.g. a random set of x4 voice sound effects, similar to the groups of sounds that are already in place)
-</user>
-<assistant>
-To help support staff assist you better, please provide the following information:
-
-- **What exactly are you trying to achieve?** Describe how you envision using the groups of custom sounds.
-- **What have you already tried?** Have you looked through the documentation or experimented with existing sound features?
-- **How would other users benefit from this feature?** What are the use cases for grouping sounds?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-**What exactly are you trying to achieve?**
-
-Groups of custom sounds, as built in like the vanilla game, so you can have randomised voice sound effects, etc
-
-**What have you already tried?**
-
-Custom sounds
-
-**How would other users benefit from this feature?**
-
-More unique and variated quests
-</user>
-<assistant>
-**What would be the advantages of this feature?** 
-
-Why should this feature be added? How would it enhance the user experience? 
-</assistant>
-<user>
-**What would be the advantages of this feature?**
-
-Players would experience a more tailored and unique quest experience by the use of custom sounds
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
