--- discord_bot/ai/intake_examples/is_there_any_way_to_interact_with_citizen_npc_only_by_right_click.md
@@ -1,16 +1,0 @@
-<user>
-# Is there any way to interact with Citizen NPC only by right-click?
-</user>
-<user>
-Hi! Due to some plugins having not been updated, my server is still staying on 1.20.1 (which means 0.4.2 Typewriter)
-
-The problem is, I have been using a different quest plugin which requires the left-click for accepting quests from an npc, yet this seems to conflict with Typewriter's either left or right click interaction.
-
-Is there any way to solve this? Like only specifying right-click to Typewriter npc interaction. (Sorry I don't know if the latest version have changed this)
-</user>
-<assistant>
-Thanks for providing such a detailed description.
-</assistant>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
