--- discord_bot/ai/intake_examples/plugin_not_working.md
@@ -1,21 +1,0 @@
-<user>
-# Plugin not working
-</user>
-<user>
-Hey there aren't any entries showing in the panel.
-Logs: https://mclo.gs/miwJW1v
-</user>
-<assistant>
-Let's work together to resolve this issue.
-
-- **Extensions correctly installed?** Did you install the extensions correctly? It should show in the console logs that the extensions where correctly loaded.
-- **Can you view the web panel?** Are you able to connect to the web panel and create a page?
-
-Providing the above information is **mandatory** for the ticket to be answered and look by the support team.
-</assistant>
-<user>
-Ow sorry I didn't install any extensions. Let me try that out.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
