--- discord_bot/ai/intake_examples/port_confusion.md
@@ -1,22 +1,0 @@
-<user>
-# Port confusion
-</user>
-<user>
-just downloaded the plugin for paper and every time i try connect to the panel it sends me to a blank page that just says "Client sent an HTTP request to an HTTPS server." and i have no clue what that means. i have made sure to change the host name and i also check that both the ports i was using were open. so now im fully out of ideas.
-</user>
-<assistant>
-To help support staff assist you effectively, please provide the following information:
-
-- **Send your config.yml** Could you please share the contents of your `config.yml` file?
-- **Check your firewall settings** Please ensure that your server's firewall allows traffic on the ports you are using. To check your firewall settings, try using [Port Checker](https://portchecker.co/) or a similar tool.
-- **Send your logs** Could you please share the contents of your `latest.log` file uploaded to [McLogs](https://mclo.gs)? 
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-I checked the firewall settings and they are open.
-Also https://mclo.gs/H5t7r
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
