--- discord_bot/ai/intake_examples/pumpkin_hat_cinematic_bug.md
@@ -1,36 +1,0 @@
-<user>
-# Pumpkin Hat Cinematic bug
-</user>
-<user>
-There is a problem when playing pumpkin hat movies after players in Minecraft put on hat equipment.
-</user>
-<assistant>
-To assist you effectively, please provide the following information:
-
-- **Logs**: Please navigate to your `logs` folder and upload the `latest.log` to [McLogs](https://mclo.gs) and share the link.
-- **Expected and Actual Behavior**: What did you expect to happen when playing the pumpkin hat movies, and what actually occurred?
-- **Reproduction Steps**: Can you outline the steps to reproduce the issue? This helps us understand how to replicate the problem.
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-https://mclo.gs/d9MV5Ck
-</user>
-<assistant>
-Thank you for sharing the logs. Now, please provide the following information to help us resolve the issue:
-
-- **Expected Behavior**: What did you expect to happen when playing the pumpkin hat movies?
-- **Actual Behavior**: What actually happened when you tried to play the movies?
-- **Reproduction Steps**: Can you outline the steps to reproduce the issue?
-</assistant>
-<user>
-After the player puts on the hat and plays the pumpkin hat movie, there is no pumpkin hat effect
-</user>
-<assistant>
-How does your segment look like for the pumpkin hat cinematic entry?
-</assistant>
-<user>
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
