--- discord_bot/ai/intake_examples/submitting_suggestions.md
@@ -1,51 +1,0 @@
-<user>
-# Submitting Suggestions
-</user>
-<user>
-What:
-I suggest shortening the time required for long pressing an entry before dragging it. In my experience, I use dragging a lot, but the long press time feels too long, which negatively affects the editing experience. Reducing this time would significantly speed up the editing process.
-
-Why:
-This adjustment would improve editing efficiency, making it easier and faster to move items around without having to wait for the long press. It would create a smoother and more responsive experience for users who frequently use the drag-and-drop function.
-
-When:
-This change would be especially useful for users who frequently engage in task editing, where quick rearrangement of entries is essential.
------------------------------------------
-What:
-I would also like to request the addition of a "Favorites" and "Merge Group" feature. This would allow users to save one or more related entries and their contents together in a group. For example, when I’m adding sound effects, particle effects, and titles for task acceptance and completion, it would be extremely helpful to save these three entries as a group. Then, I could easily access them from the "Favorites" whenever needed, saving a significant amount of time.
-
-Why:
-This feature would streamline workflows by allowing users to save and quickly access related items, reducing the need to repeatedly configure them from scratch.
-
-When:
-This would be particularly useful when managing multiple related entries for recurring tasks, especially when frequent editing and adjustments are needed.
-------------------------------------------
-Thank you for considering my suggestions! I believe these improvements would greatly enhance the overall user experience.
-<:mt:1300912896997331127> <:mh:1300912899778281566> <:acreeper:1300912873077215354> <:mn:1300912898515800187> <:mk:1300914019661774879> <:ms:1300912925107814400>
-</user>
-<assistant>
-Thank you for your detailed suggestions! To assist support staff in evaluating these requests effectively, please provide the following information for each suggestion:
-
-1. **Shortening Long Press Time:**
-   - **Use Cases:** Can you describe specific scenarios where the current long press time negatively impacts your workflow?
-   - **Expected Outcome:** What would be the ideal long press time in your opinion?
-
-2. **Adding "Favorites" and "Merge Group" Feature:**
-   - **Functionality:** Could you elaborate on how you envision the "Favorites" and "Merge Group" features working in practice?
-   - **Use Cases:** Can you provide specific examples of tasks where these features would save you time?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-Wow, Winston, you're quite smart.
-
-1.I think long pressing reduces my editing efficiency, especially when I need to copy entries in bulk. This is one of the reasons I suggested the second feature. I believe the long press mechanism could be removed and replaced with a mode where I can simply select with the left mouse button and immediately drag the entry.
-
-2.I believe I’ve been quite detailed in explaining, but when I need to edit a large number of repeated entries, the "Favorites" and "Merge Group" features would be extremely useful. For example, I could create a group named "quest_start" and add the "sound effects -> particle effects -> titles" that I mentioned earlier. Then, when I need to add the same start effects to other tasks, I could just take this group and reuse it directly.
-
-Winston, your question actually reminded me that I would like to see a "Copy Entry" feature, rather than just "link with duplicate." Right now, when I want to copy an entry, I first use "link with duplicate," then drag and unlink them, and finally drag it to the entry I want to link with. During this process, the long press drag feels awkward and choppy.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
-
