--- discord_bot/ai/intake_examples/what_was_newline_replaced_with.md
@@ -1,23 +1,0 @@
-<user>
-# What was newline replaced with
-</user>
-<user>
-All the line breaks don't work, but I can't find any new line breaks in the wiki. My NPC name needs a line feed
-</user>
-<assistant>
-To make support staff help you faster, provide the following information:
- - **What exactly are you trying to achieve?** Why do you need the line breaks for the NPC name?
- - **What have you already tried?** Have you looked through any documentation or tried using different entries?
-
-Providing the above information is **mandatory** for the ticket to be answered and look by the support team.
-</assistant>
-<user>
-I want to have a template below the npc by default. Like "black smith" or "captain". 
-
-Originally I made it with the `<newline>` in the definition and having it be two lines. But that is not possible in the latest beta anymore.
-
-I don't know how to fix it myself or even what to try...
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
