--- discord_bot/ai/intake_examples/why_is_my_staging_all_gone.md
@@ -1,23 +1,0 @@
-<user>
-# Why is my staging all gone
-</user>
-<user>
-https://mclo.gs/RUCqfuO                                                            It's normal to write opt, and when I hit publish, that's it
-</user>
-<assistant>
-To help support staff better understand the issue, please provide the following information:
-
-- **What were you trying to achieve?** Please describe what you were working on before the staging disappeared.
-- **What changes were made prior to this issue?** Did you update the plugin, change configurations, or remove any files?
-- **What is the expected outcome, and what happened instead?** What did you expect to see, and what did you actually find?
-
-Providing the above information is **mandatory** for the ticket to be answered and look by the support team.
-</assistant>
-<user>
-I was just restarting the server, before the restart, everything was fine.
-But after the restart, the staging disappeared.
-And the error in the logs occurred.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
