--- discord_bot/ai/intake_examples/wolf_entity.md
@@ -1,21 +1,0 @@
-<user>
-# Wolf Entity
-</user>
-<user>
-I tried to find wolf definition, but couldn't find it. Did I miss it, or  hasn't been added yet?
-</user>
-<assistant>
-To help support staff assist you effectively, please provide the following information:
-
-- **Did you check?** Has the wolf entity not already been added to the entity extension?
-- **Does this have priority?** Should we add the addition of this entity to our priority list?
-
-Providing the above information is **mandatory** for the ticket to be answered and looked at by the support team.
-</assistant>
-<user>
-Yeah I couldn't find it.
-Yes please add this as fast as possible.
-</user>
-<tool_call>
-complete_intake({"provided_inquiry": true})
-</tool_call>
