+++ discord_bot/src/discord/mod.rs
@@ -1,23 +1,19 @@
 mod close_ticket;
 mod create_task;
 mod post_in_questions;
-mod support_answering;
 mod task_fixed;
 mod thread_archiving;
 mod thread_cleanup;
 mod thread_closed_blocker;
-mod thread_intake;
 mod thread_support_answering;
 mod ticket_reopen;
 
 pub use close_ticket::*;
 pub use create_task::*;
 pub use post_in_questions::*;
-pub use support_answering::*;
 pub use task_fixed::*;
 pub use thread_archiving::*;
 pub use thread_cleanup::*;
 pub use thread_closed_blocker::*;
-pub use thread_intake::*;
 pub use thread_support_answering::*;
 pub use ticket_reopen::*;
